backend/requirements.txt
# Core Flask
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1

# Database + MySQL connector
mysql-connector-python==8.4.0
SQLAlchemy==2.0.32

# Environment configuration
python-dotenv==1.0.1

# AI and RAG components
requests==2.32.3
chromadb==0.5.5
sentence-transformers==3.0.1

# Utility dependencies
numpy==1.26.4
torch==2.3.0
transformers==4.41.0

# Optional (for debugging or scripts)
pytest==8.2.0
